requirements-new.txt
# Updated requirements.txt with Azure Cosmos DB support

# ============================================================
# Web Framework
# ============================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================================
# Data Validation & Settings
# ============================================================
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================================
# Azure Cosmos DB
# ============================================================
azure-cosmos==4.5.1
azure-identity==1.15.0

# ============================================================
# HTTP Clients
# ============================================================
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# ============================================================
# LangGraph & LangChain
# ============================================================
langgraph==0.0.28
langchain-core==0.1.20
langchain-openai==0.0.5

# ============================================================
# OpenAI SDK (Required)
# ============================================================
openai==1.10.0

# ============================================================
# Optional: Additional LLM Providers
# ============================================================
# Google Gemini
google-generativeai==0.3.2
langchain-google-genai==0.0.6

# Groq (fast inference)
groq==0.4.2
langchain-groq==0.0.1

# Anthropic Claude
anthropic==0.8.1

# Cohere
cohere==4.47

# Mistral
mistralai==0.0.12

# ============================================================
# Testing
# ============================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================================
# Code Quality & Linting
# ============================================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# ============================================================
# Security
# ============================================================
safety==2.3.5
bandit==1.7.6

# ============================================================
# Logging & Monitoring
# ============================================================
structlog==24.1.0
python-json-logger==2.0.7

# ============================================================
# Utilities
# ============================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
